+++ warp/src/test/java/me/sparky983/warp/internal/deserializers/BooleanDeserializerTest.java
@@ -1,6 +1,6 @@
 package me.sparky983.warp.internal.deserializers;
 
-import static me.sparky983.warp.internal.Deserializer.BOOLEAN;
+import static me.sparky983.warp.internal.Deserializers.BOOLEAN;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
