+++ warp/src/test/java/me/sparky983/warp/internal/deserializers/IntegerDeserializerTest.java
@@ -1,6 +1,6 @@
 package me.sparky983.warp.internal.deserializers;
 
-import static me.sparky983.warp.internal.Deserializer.INTEGER;
+import static me.sparky983.warp.internal.Deserializers.INTEGER;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
