+++ warp/src/test/java/me/sparky983/warp/internal/deserializers/ShortDeserializerTest.java
@@ -1,6 +1,6 @@
 package me.sparky983.warp.internal.deserializers;
 
-import static me.sparky983.warp.internal.Deserializer.SHORT;
+import static me.sparky983.warp.internal.Deserializers.SHORT;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
