+++ warp/src/test/java/me/sparky983/warp/internal/deserializers/StringDeserializerTest.java
@@ -1,6 +1,6 @@
 package me.sparky983.warp.internal.deserializers;
 
-import static me.sparky983.warp.internal.Deserializer.STRING;
+import static me.sparky983.warp.internal.Deserializers.STRING;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
